feat: v0.8.0 coordinate transform mode, operation capabilities, split large operation files
